+++ FISP_git/src/com/synesoft/ftzmis/app/controller/FTZ210311Controller.java
@@ -21,9 +21,9 @@ import com.synesoft.fisp.app.common.utils.StringUtil;
 import com.synesoft.fisp.domain.model.UserInf;
 import com.synesoft.fisp.domain.service.NumberService;
 import com.synesoft.ftzmis.app.common.constants.CommonConst;
+import com.synesoft.ftzmis.app.common.msgproc.FtzMsgHead;
+import com.synesoft.ftzmis.app.common.msgproc.FtzMsgProcService;
 import com.synesoft.ftzmis.app.common.util.DateUtil;
-import com.synesoft.ftzmis.app.common.xmlproc.GenerateXml;
-import com.synesoft.ftzmis.app.common.xmlproc.MsgHead;
 import com.synesoft.ftzmis.app.model.FTZ210311Form;
 import com.synesoft.ftzmis.domain.model.FtzBankCode;
 import com.synesoft.ftzmis.domain.model.FtzOffMsgCtl;
@@ -48,7 +48,7 @@ public class FTZ210311Controller {
 	private NumberService numberService;
 	
 	@Resource
-	protected GenerateXml generateXml;
+	protected FtzMsgProcService generateXml;
 	
 	@ModelAttribute
 	public FTZ210311Form setForm() {
@@ -136,7 +136,7 @@ public class FTZ210311Controller {
 				.getFormatDateRemoveSprit(insert_FtzOffMsgCtl.getWorkDate()));
 		
 		// 设置批量头信息
-		MsgHead mh = MsgHead.getMsgHead();
+		FtzMsgHead mh = FtzMsgHead.getMsgHead();
 		insert_FtzOffMsgCtl.setVer(mh.getVER());
 		insert_FtzOffMsgCtl.setSrc(mh.getSRC());
 		insert_FtzOffMsgCtl.setDes(mh.getDES());
@@ -663,7 +663,7 @@ public class FTZ210311Controller {
 					model.addAttribute(ResultMessages.error().add(
 					"e.ftzmis.210101.0035"));
 					form.setSelected_msgId("");
-					return "forward:/FTZ210310/AddQry";
+					return "forward:/FTZ210311/AddQry";
 				}
 			}
 		}
@@ -699,7 +699,16 @@ public class FTZ210311Controller {
 		FtzOffTxnDtl del_FtzOffTxnDtl = new FtzOffTxnDtl();
 		del_FtzOffTxnDtl.setMsgId(form.getSelected_msgId());
 		del_FtzOffTxnDtl.setSeqNo(form.getSelected_seqNo());
-
+		
+		FtzOffTxnDtl result_FtzOffTxnDtl = ftz210310Service.queryFtzOffTxnDtl(del_FtzOffTxnDtl);
+		if (CommonConst.FTZ_MSG_STATUS_AUTH_SUCC.equals(result_FtzOffTxnDtl.getChkStatus())) {
+			model.addAttribute(ResultMessages.error().add(
+			"e.ftzmis.210101.0037"));
+			form.setSelected_msgId("");
+			form.setSelected_seqNo(null);
+			return "forward:/FTZ210311/UpdDtlInit";
+		}
+		
 		int i = ftz210310Service.deleteFtzOffTxnDtl(del_FtzOffTxnDtl);
 
 		if (i < 0) {
@@ -934,16 +943,18 @@ public class FTZ210311Controller {
 			}
 			if (count_unAuth > 0) {
 				model.addAttribute(ResultMessages.error().add(
-						"e.ftzmis.210101.0024",
+						"e.ftzmis.210310.0012",
 						sb_unAuth.subSequence(0, sb_unAuth.length() - 1)));
+				form.setAuthFinishFlag("");
 				return "forward:/FTZ210311/QryAuthDtl";
 			}
 			
 			
 			if (count_authFail > 0) {
 				model.addAttribute(ResultMessages.error().add(
-						"e.ftzmis.210101.0031",
+						"e.ftzmis.210310.0012",
 						sb_unAuth.subSequence(0, sb_authFail.length() - 1)));
+				form.setAuthFinishFlag("");
 				return "forward:/FTZ210311/QryAuthDtl";
 			} 
 			
@@ -963,7 +974,7 @@ public class FTZ210311Controller {
 			} else {
 				if (update_FtzOffMsgCtl.getMsgStatus().equals(
 						CommonConst.FTZ_MSG_STATUS_AUTH_SUCC)) {
-					generateXml.writeXml(update_FtzOffMsgCtl.getMsgNo(),
+					generateXml.submitMsg(update_FtzOffMsgCtl.getMsgNo(),
 							update_FtzOffMsgCtl.getMsgId());
 				}
 				model.addAttribute(ResultMessages.success().add(
